fast_api_backend/requirements.txt
fastapi==0.111.0
openai==0.28.0
python-dotenv==1.0.1
sentence_transformers==2.2.2
uvicorn==0.30.1
weaviate_client==4.5.0
